requirements.txt
streamlit
pandas
openpyxl
sqlalchemy
faker
fpdf2
streamlit-cookies-controller